Factor out lock indexing test utility.
