+++ tests/integration/test_lock_resolver.py
@@ -20,12 +20,11 @@ from pex.interpreter import PythonInterpreter
 from pex.pep_440 import Version
 from pex.pep_503 import ProjectName
 from pex.pex_info import PexInfo
-from pex.resolve.locked_resolve import LockedRequirement
-from pex.resolve.lockfile import json_codec
 from pex.typing import TYPE_CHECKING
 from pex.util import CacheHelper
 from testing import IS_PYPY, PY_VER, built_wheel, make_env, run_pex_command
 from testing.cli import run_pex3
+from testing.lock import index_lock_artifacts
 
 if TYPE_CHECKING:
     from typing import Any, Mapping, Tuple
@@ -49,18 +48,6 @@ def index_pex_distributions(pex_file):
     return dict(
         project_name_and_version(location) for location in PexInfo.from_pex(pex_file).distributions
     )
-
-
-def index_lock_artifacts(lock_file):
-    # type: (str) -> Mapping[ProjectName, LockedRequirement]
-
-    lock = json_codec.load(lock_file)
-    assert 1 == len(lock.locked_resolves)
-    locked_resolve = lock.locked_resolves[0]
-    return {
-        locked_requirement.pin.project_name: locked_requirement
-        for locked_requirement in locked_resolve.locked_requirements
-    }
 
 
 @pytest.fixture(scope="module")
